adds TableRowExpandable
