+++ frontend/src/components/UI/TableRowExpandable/index.js
@@ -1,0 +1,1 @@
+export { default } from "./TableRowExpandable";
